+++ internal/models/notification.model.go
@@ -1,0 +1,7 @@
+package models
+
+type Notification struct {
+	ID    int    `json:"id"`
+	Key   string `json:"key"`
+	Value string `json:"value"`
+}