+++ internal/router/farm/farm.router.go
@@ -1,0 +1,11 @@
+package farm
+
+import (
+	"avatar/internal/controller"
+
+	"github.com/gin-gonic/gin"
+)
+
+func FarmRoutes(r *gin.RouterGroup) {
+	r.GET("/farm-data", controller.GetFarmData)
+}