+++ internal/router/notification/notification.router.go
@@ -1,0 +1,11 @@
+package notification
+
+import (
+	"avatar/internal/controller"
+
+	"github.com/gin-gonic/gin"
+)
+
+func NotificationRoutes(r *gin.RouterGroup) {
+	r.POST("/notification", controller.GetNotification)
+}