+++ CHANGELOG.md
@@ -7,4 +7,6 @@
 
 ### [30/03/2025](https://github.com/vungnv10103/ServerAvatar258)
 
+- [avatar/22:25](https://github.com/vungnv10103/ServerAvatar258/issues) add `notification`, `farm data`.
+
 - [avatar/18:10](https://github.com/vungnv10103/ServerAvatar258/issues) init project.